+++ docs/changelog.md
@@ -19,6 +19,7 @@ MML changes:
  * Fixed missing `set_detune 0` instruction when `MD` value is small
  * Fixed MP and portamento not accounting for manual `D` detune
  * Fixed portamento not testing if detune was changed after the previous note
+ * Added instrument tracking for previous slurred note before portamento
 
 GUI changes:
  * Sample Analyser can now analyse samples up to a length of 32768.